Add TableOfContentsSidebar to legal notices
