+++ src/components/doc/index.js
@@ -1,2 +1,1 @@
 export { default as Sidebar } from './Sidebar';
-export { default as TableOfContentsSidebar } from './TableOfContentsSidebar';
