+++ src/templates/LegalNotice.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import { createUseStyles } from 'react-jss';
 import { SEO, StickyFooter, ContentHeader } from 'components';
-
+import { TableOfContentsSidebar } from 'components/Sidebar';
 import { Sidebar } from 'components/legal-notice';
 
 const useStyles = createUseStyles((theme) => ({
@@ -16,6 +16,7 @@ const useStyles = createUseStyles((theme) => ({
     article: {
       paddingTop: 32,
       paddingLeft: 32,
+      paddingRight: 32,
     },
 
     main: {
@@ -47,6 +48,8 @@ const LegalNotice = ({ data: { notice, site }, location }) => {
               dangerouslySetInnerHTML={{ __html: notice.html }}
             />
           </article>
+
+          <TableOfContentsSidebar headings={notice.headings} />
         </main>
       </StickyFooter>
     </>
@@ -69,6 +72,10 @@ export const pageQuery = graphql`
     notice: markdownRemark(fields: { slug: { eq: $slug } }) {
       id
       html
+      headings(depth: h3) {
+        id
+        value
+      }
       frontmatter {
         description
         title
